Move model loading to model sub-package's init file
